objects menu done
the menu is saved as a prefab so it should be possible to use it between other projects. it requires another prefab called insertObjects, where you put any objects you want. Menu will show up when the left touchpad is pressed and then you need to hold down the touchpad to select the object you want and then press the trigger to make the selection.